Update code to account for uint* mapping change
This updates application logic to account for the change of
how uint{8,16,32,64} types are mapped.

+++ opendaylight/md-sal/samples/clustering-test-app/provider/src/main/java/org/opendaylight/controller/clustering/it/provider/MdsalLowLevelTestProvider.java
@@ -258,7 +258,7 @@ public class MdsalLowLevelTestProvider implements OdlMdsalLowlevelControlService
         LOG.info("In startPublishNotifications - input: {}", input);
 
         final PublishNotificationsTask task = new PublishNotificationsTask(notificationPublishService, input.getId(),
-                input.getSeconds(), input.getNotificationsPerSecond());
+                input.getSeconds().longValue(), input.getNotificationsPerSecond().longValue());
 
         publishNotificationsTasks.put(input.getId(), task);
